+++ src/verify/table_event_app.sql
@@ -1,0 +1,10 @@
+BEGIN;
+
+SELECT id,
+       app_id,
+       event_id,
+       created_at,
+       created_by
+FROM vibetype.event_app WHERE FALSE;
+
+ROLLBACK;